Add WithStatus  (which I had before) back
But in a way that also keeps Matthias happy because the MakesResource
machinery stays in tact

If you chose mkres = I and  res = WithStatus n a, then it is exactly
the same as what I had before Matthias came up with this abstraction.

I think it's a bit easier to use when users don't define their own
Resource type.

We should experiment with both approaches and see which one is
friendlier. This will just require some experience reports from
us using this library at Wire